Refactor Playwright tools actions into grouped partial classes
